requirements.txt
langchain>=1.2.13
langchain-core>=1.2.24
langchain-openai>=1.1.12
langchain-community>=0.4.1
langchain-classic>=1.0.3
langgraph>=1.1.0
faiss-cpu
rank_bm25
sentence-transformers
ddgs>=9.12.0
trafilatura>=2.0.0
pypdf
pydantic>=2.12.5
pydantic-settings>=2.12.5
